extract `State` to its own file
